+++ packages/components/input/src/InputAddon.styles.ts
@@ -1,12 +1,31 @@
 import { cva, VariantProps } from 'class-variance-authority'
 
-export const inputAddonStyles = cva(['flex', 'items-center'], {
-  variants: {
-    isDisabled: {
-      true: ['bg-on-surface/dim-5', 'text-on-surface/dim-3', 'cursor-not-allowed'],
-      false: ['bg-surface', 'text-on-surface'],
+export const inputAddonStyles = cva(
+  ['ring-1 ring-inset', 'shrink-0', 'focus-visible:relative', 'focus-visible:z-raised'],
+  {
+    variants: {
+      asChild: { false: ['flex', 'items-center', 'px-lg'] },
+      intent: {
+        neutral: 'ring-outline',
+        error: 'ring-error',
+        alert: 'ring-alert',
+        success: 'ring-success',
+      },
+      disabled: {
+        true: ['pointer-events-none'],
+      },
     },
-  },
-})
+    compoundVariants: [
+      {
+        disabled: false,
+        asChild: false,
+        class: ['bg-surface', 'text-on-surface'],
+      },
+    ],
+    defaultVariants: {
+      intent: 'neutral',
+    },
+  }
+)
 
 export type InputAddonStylesProps = VariantProps<typeof inputAddonStyles>