+++ packages/components/input/src/InputGroupContext.ts
@@ -1,17 +1,19 @@
 import { createContext, useContext } from 'react'
 
-import { InputContainerProps } from './InputContainer'
-
-export interface InputGroupContext extends Pick<InputContainerProps, 'intent'> {
-  isDisabled?: boolean
-  isLeftElementVisible: boolean
-  isRightElementVisible: boolean
-  isLeftAddonVisible: boolean
-  isRightAddonVisible: boolean
+export interface InputGroupContext {
+  disabled?: boolean
+  hasLeadingIcon: boolean
+  hasTrailingIcon: boolean
+  hasLeadingAddon: boolean
+  hasTrailingAddon: boolean
+  state: null | undefined | 'error' | 'alert' | 'success'
+  isStandalone?: boolean
 }
 
 export const InputGroupContext = createContext<Partial<InputGroupContext> | null>(null)
 
 export const useInputGroup = () => {
-  return useContext(InputGroupContext)
+  const context = useContext(InputGroupContext)
+
+  return context || { isStandalone: true }
 }