+++ packages/components/radio-group/src/Radio.tsx
@@ -13,15 +13,14 @@ export const Radio = forwardRef<HTMLButtonElement, RadioProps>(
     const innerId = useId()
     const innerLabelId = useId()
 
-    const { intent, size, disabled } = useRadioGroup()
+    const { intent, disabled } = useRadioGroup()
 
     return (
-      <div className={cx('flex items-center gap-md text-body-1', className)}>
+      <div className={cx('flex items-start gap-md text-body-1', className)}>
         <RadioInput
           ref={ref}
           id={id || innerId}
           intent={intent}
-          size={size}
           aria-labelledby={children ? innerLabelId : undefined}
           {...others}
         />