+++ packages/components/icon-button/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.9](https://github.com/adevinta/spark/compare/@spark-ui/icon-button@1.3.8...@spark-ui/icon-button@1.3.9) (2023-07-10)
+
+**Note:** Version bump only for package @spark-ui/icon-button
+
+## [1.3.8](https://github.com/adevinta/spark/compare/@spark-ui/icon-button@1.3.7...@spark-ui/icon-button@1.3.8) (2023-07-10)
+
+**Note:** Version bump only for package @spark-ui/icon-button
+
+## [1.3.7](https://github.com/adevinta/spark/compare/@spark-ui/icon-button@1.3.6...@spark-ui/icon-button@1.3.7) (2023-07-05)
+
+**Note:** Version bump only for package @spark-ui/icon-button
+
 ## [1.3.6](https://github.com/adevinta/spark/compare/@spark-ui/icon-button@1.3.5...@spark-ui/icon-button@1.3.6) (2023-06-28)
 
 **Note:** Version bump only for package @spark-ui/icon-button