+++ packages/components/icons/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.19.1](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.19.0...@spark-ui/icons@1.19.1) (2023-07-13)
+
+### Bug Fixes
+
+- **icons:** remove flag icons ([4e43fd3](https://github.com/adevinta/spark/commit/4e43fd35dffca6746804bdc918a28f2c17772361))
+
+### Reverts
+
+- **icons:** handle country flag colors ([93e45ff](https://github.com/adevinta/spark/commit/93e45ffe21401bb087726c60672c68760861e38e))
+
+# [1.19.0](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.18.0...@spark-ui/icons@1.19.0) (2023-07-04)
+
+### Features
+
+- **icons:** add country flag icons ([27d4fb9](https://github.com/adevinta/spark/commit/27d4fb9c7a7ed5b449e3593f8c772b45cb4da726))
+- **icons:** handle country flag colors ([e67dd6f](https://github.com/adevinta/spark/commit/e67dd6ff7c4dd54560b1f482823669a4f912fcbc))
+
 # [1.18.0](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.17.0...@spark-ui/icons@1.18.0) (2023-06-30)
 
 ### Bug Fixes